# MODULO1_ENTREGABLE_COMPLETO/openapi/module1_openapi.yaml
openapi: 3.0.3
info:
  title: Love to Help – API Módulo 1
  version: 1.0.0
  description: >
    Endpoints base de autenticación, usuario y sistema para el Módulo 1.
servers:
  - url: https://api.lovetohelp.com
paths:
  /auth/register:
    post:
      summary: Registro de nuevo usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, country]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                country:
                  type: string
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Datos inválidos
        '409':
          description: Email ya registrado

  /auth/login:
    post:
      summary: Login de usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Credenciales inválidas

  /auth/refresh:
    post:
      summary: Refrescar tokens
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Refresh token inválido

  /auth/logout:
    post:
      summary: Logout
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout exitoso

  /users/me:
    get:
      summary: Obtener el usuario actual
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: No autenticado

  /users/me/profile:
    get:
      summary: Obtener el perfil del usuario actual
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    patch:
      summary: Actualizar el perfil del usuario actual
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  /system/health:
    get:
      summary: Healthcheck del sistema
      tags: [System]
      responses:
        '200':
          description: Sistema operativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  time:
                    type: string
                    format: date-time

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        status:
          type: string
        country:
          type: string

    UserProfile:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        birthdate:
          type: string
          format: date
        address:
          type: object
          additionalProperties: true
        documentType:
          type: string
        documentNumber:
          type: string
        profilePhotoUrl:
          type: string

    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthdate:
          type: string
          format: date
        address:
          type: object
          additionalProperties: true
        documentType:
          type: string
        documentNumber:
          type: string
        profilePhotoUrl:
          type: string

    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string